Updated directory output paths
The output directories have been updated but tests have not been. Additionally the output folder structure is modifiable within the modules.config file at the top level to fit in with the other pipelines, I think it is a cleaner way of specifying things as well. Tests have not been updated yet, however they will be after the module output names have been updated too reflect the process used to run them. As the tests will have to be updated twice otherwise.

Additionally, the length filtered contigs are now set as the output contigs for follow up processes as they were not previously