removed annotate from locidex search parameters
If locidex receives a fasta that has been processed with extract it no longer needs to be annotated with pyrodigal as the sample has been blasted against the database to extract relevent hits. To verify that results generated in locidex match what is tested, this is the preferred route forward. However as databases will get larger in may be more cost effective pass in a genbank file or just annotate the fasta directly, however this relies on bakta working...